Update problems to use path.extname over RegExp.
While regex traps like this are definitely essential to know, it seems
to cause much confusion and really isn't crucial to people's learning of
node.

path.extname simplifies this business a whole bunch and is probably
better practice anyway; it says exactly what it does and avoids the
possibility your regex is harboring "gotchas".

https://github.com/nodeschool/discussions/issues/91
https://github.com/nodeschool/discussions/issues/112