Add extra hint pointer for command arguments
A few folks have gotten stuck briefly[1] on process.argv[2] in the IO exercises. Since that's not really the focus of the tutorial past (Baby Steps), overtly and briefly hint that the answer is in the previous tutorial. 

There should be no need to repeat this hint later or paste the entire hint from Baby Steps here.

hth,
adric