Added explicit instruction about test file
I've seen this twice: a beginner makes their own test file to test their
code that checks the line length of a file, and then is confused when
the tests fail. I think this line is a bit confusing - it would be better
to add in something that tells the user they don't need to make their
own file, just to help readers who don't read this line too much.