Clarification
Here I'm trying to make it more explicit that a global variable is needed as a container for the sum. I didn't pick up on this and tried to fit everything into the `for loop`. I might be spelling things out too much with this commit, but I had to look at the solution before I could understand this problem. 

I also tried to make the writing in this problem more concise.