Update problem.md
The folder is passed as second argument. The extension array list is passed as third argument.
The documentation is confusing. As, it is not clearly saying the arguments position.
Because, i tried several times. Before checking the solution. To find out that my input argument are for folder is 1 and extension list is 2.
Here is my code
```
var fs = require('fs');
var path = require('path');

var bufferObjectOfTheFile;

function bufferTheFile(pathToTheFile) {
  fs.readdir(pathToTheFile, function (err, list) {
	// If there is an error. It will print out the error.
	list.forEach(function (item) {
		if(path.extname(item).replace('.','') === process.argv[2]) {
			console.log(item);
		}
	});
	if(err) {
		console.log(err);
	}
} );
}

bufferTheFile(process.argv[1]);
```
Upon, changing the arguments from 1 to 2 and 2 to 3. The code is passing the test.