Update README.md
The order of the command to install learnyounode was not working, so I realize that the default way to do that is to put the global command "-g" before "learnyounode". I don't no if this was made on purpose, but for me, someone that is trying to learn by using learnyounode on a local environment, that way was not working.